Refactor some folder structure
