+++ src/common/Util.ts
@@ -1,0 +1,3 @@
+export function sleep (ms: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
