+++ src/common/s3/NoopRequestSigner.ts
