+++ src/deal-preparation/DealPreparationWorker.ts
@@ -6,7 +6,7 @@ import Datastore from '../common/Datastore';
 import { Category } from '../common/Logger';
 import GenerationRequest from '../common/model/GenerationRequest';
 import ScanningRequest from '../common/model/ScanningRequest';
-import Scanner from './Scanner';
+import Scanner from './scanner/Scanner';
 import fs from 'fs-extra';
 import path from 'path';
 import { performance } from 'perf_hooks';
@@ -15,9 +15,9 @@ import { FileInfo, FileList } from '../common/model/InputFileList';
 import GenerateCar from '../common/GenerateCar';
 import { pipeline } from 'stream/promises';
 import { S3Client, GetObjectCommand, GetObjectCommandInput } from '@aws-sdk/client-s3';
-import NoopRequestSigner from './NoopRequestSigner';
+import NoopRequestSigner from '../common/s3/NoopRequestSigner';
 import winston from 'winston';
-import { getRetryStrategy } from '../common/S3RetryStrategy';
+import { getRetryStrategy } from '../common/s3/S3RetryStrategy';
 import pAll from 'p-all';
 import * as stream from 'stream';
 import { TransformCallback } from 'stream';
